+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added the fast-forward & rewind buttons for the Android notification when `mediaControl.mediaSessionEnabled` is set to `true`.
 
+### Fixed
+
+- Fixed an issue on Android where the notification would not disappear when setting an undefined source.
+
+### Changed
+
+- Replaced the `MediaBrowserService` with a regular `Service` to facilitate background playback on Android.
+
 ## [7.7.1] - 24-07-29
 
 ### Fixed